update modal with new file pattern
